change float to BigDecimal

+++ backend/src/main/java/com/fairpay/currency/vo/RelativeCurrency.java
@@ -2,15 +2,17 @@ package com.fairpay.currency.vo;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class RelativeCurrency {
   @JsonProperty("price")
-  protected float price;
+  protected BigDecimal price;
 
-  public float getPrice() {
+  public BigDecimal getPrice() {
     return price;
   }
 
-  public void setPrice(float price) {
+  public void setPrice(BigDecimal price) {
     this.price = price;
   }
